requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.1
Django==3.2.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
